ref: move mongo repositories to db directory
