--- data/index.ts
@@ -1,4 +1,0 @@
-import MongoBudgetRepository from './budget/MongoBudgetRespository';
-import MongoExpenseRepository from './expense/MongoExpenseRepository';
-
-export { MongoBudgetRepository, MongoExpenseRepository };
