+++ db/mongo/index.ts
@@ -2,6 +2,8 @@ import mongoose from 'mongoose';
 
 import budgetModel from './budget';
 import expenseModel from './expense';
+import MongoBudgetRepository from './budget/MongoBudgetRespository';
+import MongoExpenseRepository from './expense/MongoExpenseRepository';
 import migrate from './migration';
 
 async function connectMongo() {
@@ -19,4 +21,4 @@ function disconnectMongo() {
   mongoose.disconnect();
 }
 
-export { connectMongo as connectDb, disconnectMongo as disconnectDb, budgetModel, expenseModel };
+export { connectMongo as connectDb, disconnectMongo as disconnectDb, budgetModel, expenseModel, MongoBudgetRepository, MongoExpenseRepository };
